requirements.txt
requests==2.31.0
pytest==7.3.2
pydantic==2.0
structlog==23.1.0
curlify==2.2.1
PyHamcrest==2.0.4
allure-pytest==2.13.2
allure-python-commons==2.13.2